Proof-of-concept Cilium integration for openshift-ansible.
Tested with OKD 3.11, using master branch of openshift-ansible from 3-2-19

Many of the hacks unrelated to Cilium were workarounds that seem to have
been required due to the fact that I was testing on RHEL 8 beta.  I
suspect that if you ran the same Cilium ansible playbook with Fedora 28/29
these hacks would not be needed.

This installation of Cilium allowed me to have all of the openhift infrastructure
pods come up in a good state, but DNS was not working for regular pods when
I created a test app.  I suspect this is related to differences in how openshift
does DNS as part of the 'proxy-and-dns' daemonset that runs on each host.
I also needed to add the pod-label 'k8s-app=kube-dns' to the pod spec in that
daemonset definition.   This allows cilium to let this pod start up even before
Cilium has been able to talk to etcd or the k8s api server.

The configuration options for Cilium is pretty limited.   Adding the following
to your inventory.ini should be sufficient:

os_sdn_network_plugin_name=cni
openshift_use_openshift_sdn=false
openshift_use_cilium=true

Current setup installs the default Cilium YAML, but with cilium-etcd-operator
disabled in favor of using the etcd that is part of the master node.