Proof-of-concept Cilium integration for openshift-ansible.
Tested with OKD 3.11, using master branch of openshift-ansible from 3-2-19
in a single-node openshift environment.

This installation of Cilium allowed me to have all of the openshift infrastructure
pods come up in a good state.  However, the environment is not yet functional for
at least two reasons:

(1) DNS was not working for regular pods when I created a test app.
I suspect this is related to differences in how openshift
does DNS as part of the 'proxy-and-dns' daemonset that runs on each host.
(2) OKD does not set the standard node annotations used by Cilium get the
the node-cidr, and as a result, Cilium IPAM allocates pod IPs from a range that is
different from what Openshift components (namely the proxy-and-dns) pods think
is the cluster cidr range.   This is likely to cause problems with things like NodePort,
and possibly is related to the above DNS issue.

The configuration options for Cilium is pretty limited.   Current setup installs the default
Cilium YAML, but using the etcd that is part of the master node rather than etcd-operator.
Adding the following to your inventory.ini should be sufficient:

os_sdn_network_plugin_name=cni
openshift_use_openshift_sdn=false
openshift_use_cilium=true